refactor: 지도 검색 조회와 geocoding 책임 분리
- CollectionSpotRepository를 raw search 전용 interface로 정리
- CollectionSpotGeocodingRepository 추가
- CollectionSpotRepositoryImpl과 CollectionSpotGeocodingRepositoryImpl 분리
- keyword 검색의 raw search -> region filter -> geocode 흐름 유지
- location/current map 검색을 raw search -> geocode 흐름으로 통일
- data 테스트를 검색 조회와 geocoding 책임별로 분리
- location 검색 파이프라인 유지 테스트 추가
- GEOCODING_CONCURRENCY_LIMIT = 3 유지
