+++ data/src/test/java/com/team/yeogibeoryeo/data/spot/repository/CollectionSpotRepositoryImplTest.kt
@@ -1,7 +1,6 @@
 package com.team.yeogibeoryeo.data.spot.repository
 
 import com.team.yeogibeoryeo.data.core.key.AppKeyProvider
-import com.team.yeogibeoryeo.data.spot.geocoder.SpotGeocoder
 import com.team.yeogibeoryeo.data.spot.mapper.SpotMapper
 import com.team.yeogibeoryeo.data.spot.remote.SpotApiService
 import com.team.yeogibeoryeo.data.spot.remote.datasource.SpotRemoteDataSource
@@ -16,22 +15,21 @@ import com.team.yeogibeoryeo.domain.spot.model.Coordinate
 import kotlinx.coroutines.runBlocking
 import kotlinx.serialization.json.JsonPrimitive
 import org.junit.Assert.assertEquals
-import org.junit.Assert.assertNotNull
 import org.junit.Assert.assertNull
 import org.junit.Test
 
 class CollectionSpotRepositoryImplTest {
 
     @Test
-    fun `검색어 기반 조회 결과를 CollectionSpot으로 변환하고 좌표를 설정한다`() = runBlocking {
+    fun `검색어 기반 raw 조회 결과를 CollectionSpot으로 변환한다`() = runBlocking {
         val apiService = FakeSpotApiService(
             response = createNormalResponse(),
         )
         val repository = createRepository(apiService)
 
-        val result = repository.searchByKeyword(
+        val result = repository.searchRawByKeyword(
             keyword = "문래동",
-        )
+        ).spots
 
         assertEquals(TEST_SERVICE_KEY, apiService.requestedServiceKey)
         assertEquals("문래동", apiService.requestedAddr)
@@ -43,9 +41,7 @@ class CollectionSpotRepositoryImplTest {
         assertEquals("서울특별시 영등포구 문래동", firstSpot.address)
         assertEquals("주민센터 앞", firstSpot.detailLocation)
         assertEquals(CollectionSpotType.BATTERY_BIN, firstSpot.type)
-        assertNotNull(firstSpot.coordinate)
-        assertEquals(37.5, firstSpot.coordinate?.latitude)
-        assertEquals(126.9, firstSpot.coordinate?.longitude)
+        assertNull(firstSpot.coordinate)
     }
 
     @Test
@@ -55,10 +51,10 @@ class CollectionSpotRepositoryImplTest {
         )
         val repository = createRepository(apiService)
 
-        val result = repository.searchByKeyword(
+        val result = repository.searchRawByKeyword(
             keyword = "문래동",
             types = setOf(CollectionSpotType.RECYCLING_CENTER),
-        )
+        ).spots
 
         assertEquals(TEST_SERVICE_KEY, apiService.requestedServiceKey)
         assertEquals(1, result.size)
@@ -71,20 +67,17 @@ class CollectionSpotRepositoryImplTest {
         val apiService = FakeSpotApiService(
             response = createNormalResponse(),
         )
-        val spotGeocoder = FakeSpotGeocoder()
         val repository = createRepository(
             apiService = apiService,
-            spotGeocoder = spotGeocoder,
         )
 
-        val result = repository.searchByKeywordResultWithoutCoordinates(
+        val result = repository.searchRawByKeyword(
             keyword = "문래동",
         )
 
         assertEquals(TEST_SERVICE_KEY, apiService.requestedServiceKey)
         assertEquals("문래동", apiService.requestedAddr)
         assertEquals(2, result.spots.size)
-        assertEquals(emptyList<String>(), spotGeocoder.requestedAddresses)
         assertNull(result.spots.first().coordinate)
     }
 
@@ -113,9 +106,9 @@ class CollectionSpotRepositoryImplTest {
         )
         val repository = createRepository(apiService)
 
-        val result = repository.searchByKeyword(
+        val result = repository.searchRawByKeyword(
             keyword = "문래동",
-        )
+        ).spots
 
         assertEquals(listOf(1, 2), apiService.requestedPageNos)
         assertEquals(
@@ -131,7 +124,7 @@ class CollectionSpotRepositoryImplTest {
         )
         val repository = createRepository(apiService)
 
-        val result = repository.searchByLocation(
+        val result = repository.searchRawByLocation(
             coordinate = Coordinate(
                 latitude = 37.5182396969791,
                 longitude = 126.895880210522,
@@ -154,147 +147,20 @@ class CollectionSpotRepositoryImplTest {
         )
         val repository = createRepository(apiService)
 
-        val result = repository.searchByKeyword(
+        val result = repository.searchRawByKeyword(
             keyword = "약수동",
-        )
+        ).spots
 
         assertEquals(TEST_SERVICE_KEY, apiService.requestedServiceKey)
         assertEquals(emptyList<Any>(), result)
     }
 
-    @Test
-    fun `같은 주소는 한 번만 geocoding하고 기존 순서를 유지한다`() = runBlocking {
-        val apiService = FakeSpotApiService(
-            response = createDuplicateAddressResponse(),
-        )
-        val spotGeocoder = FakeSpotGeocoder(
-            coordinatesByAddress = mapOf(
-                "서울특별시 영등포구 문래동" to Coordinate(
-                    latitude = 37.5,
-                    longitude = 126.9,
-                ),
-                "서울특별시 구로구 구로동" to Coordinate(
-                    latitude = 37.4,
-                    longitude = 126.8,
-                ),
-            ),
-        )
-        val repository = createRepository(
-            apiService = apiService,
-            spotGeocoder = spotGeocoder,
-        )
-
-        val result = repository.searchByKeyword(
-            keyword = "문래동",
-        )
-
-        assertEquals(
-            listOf(
-                "폐건전지 수거함",
-                "폐건전지 수거함 2",
-                "재활용센터",
-            ),
-            result.map { it.name },
-        )
-        assertEquals(
-            listOf(
-                "서울특별시 영등포구 문래동",
-                "서울특별시 구로구 구로동",
-            ),
-            spotGeocoder.requestedAddresses,
-        )
-        assertEquals(result[0].coordinate, result[1].coordinate)
-        assertEquals(37.4, result[2].coordinate?.latitude)
-    }
-
-    @Test
-    fun `괄호가 포함된 도로명 주소는 괄호를 제거한 주소로 geocoding한다`() = runBlocking {
-        val apiService = FakeSpotApiService(
-            response = createParenthesizedRoadAddressResponse(),
-        )
-        val spotGeocoder = FakeSpotGeocoder(
-            coordinatesByAddress = mapOf(
-                "서울특별시 성동구 행당로 35" to Coordinate(
-                    latitude = 37.5570,
-                    longitude = 127.0332,
-                ),
-            ),
-        )
-        val repository = createRepository(
-            apiService = apiService,
-            spotGeocoder = spotGeocoder,
-        )
-
-        val result = repository.searchByKeyword(
-            keyword = "금호동",
-        )
-
-        assertEquals(
-            listOf("서울특별시 성동구 행당로 35"),
-            spotGeocoder.requestedAddresses,
-        )
-        assertEquals(37.5570, result.first().coordinate?.latitude)
-        assertEquals(127.0332, result.first().coordinate?.longitude)
-    }
-
-    @Test
-    fun `공백 주소는 geocoding하지 않고 좌표 없이 유지한다`() = runBlocking {
-        val apiService = FakeSpotApiService(
-            response = createBlankAddressResponse(),
-        )
-        val spotGeocoder = FakeSpotGeocoder()
-        val repository = createRepository(
-            apiService = apiService,
-            spotGeocoder = spotGeocoder,
-        )
-
-        val result = repository.searchByKeyword(
-            keyword = "문래동",
-        )
-
-        assertEquals(1, result.size)
-        assertNull(result.first().coordinate)
-        assertEquals(emptyList<String>(), spotGeocoder.requestedAddresses)
-    }
-
-    @Test
-    fun `일부 geocoding 실패가 전체 검색 결과를 실패시키지 않는다`() = runBlocking {
-        val apiService = FakeSpotApiService(
-            response = createNormalResponse(),
-        )
-        val spotGeocoder = FakeSpotGeocoder(
-            coordinatesByAddress = mapOf(
-                "서울특별시 구로구 구로동" to Coordinate(
-                    latitude = 37.4,
-                    longitude = 126.8,
-                ),
-            ),
-            failureAddresses = setOf("서울특별시 영등포구 문래동"),
-        )
-        val repository = createRepository(
-            apiService = apiService,
-            spotGeocoder = spotGeocoder,
-        )
-
-        val result = repository.searchByKeyword(
-            keyword = "문래동",
-        )
-
-        assertEquals(2, result.size)
-        assertEquals("폐건전지 수거함", result[0].name)
-        assertNull(result[0].coordinate)
-        assertEquals("재활용센터", result[1].name)
-        assertEquals(37.4, result[1].coordinate?.latitude)
-    }
-
     private fun createRepository(
         apiService: FakeSpotApiService,
-        spotGeocoder: SpotGeocoder = FakeSpotGeocoder(),
     ): CollectionSpotRepositoryImpl {
         return CollectionSpotRepositoryImpl(
             remoteDataSource = SpotRemoteDataSource(apiService),
             spotMapper = SpotMapper(),
-            spotGeocoder = spotGeocoder,
             publicDataKeyProvider = FakePublicDataKeyProvider(),
         )
     }
@@ -302,27 +168,6 @@ class CollectionSpotRepositoryImplTest {
     private class FakePublicDataKeyProvider : AppKeyProvider {
         override val publicDataServiceKey: String = TEST_SERVICE_KEY
         override val naverClientId: String = "naver-client-id"
-    }
-
-    private class FakeSpotGeocoder(
-        private val coordinatesByAddress: Map<String, Coordinate> = emptyMap(),
-        private val failureAddresses: Set<String> = emptySet(),
-    ) : SpotGeocoder {
-
-        val requestedAddresses = mutableListOf<String>()
-
-        override suspend fun geocode(address: String): Coordinate? {
-            requestedAddresses += address
-
-            if (address in failureAddresses) {
-                error("geocoding failed")
-            }
-
-            return coordinatesByAddress[address] ?: Coordinate(
-                latitude = 37.5,
-                longitude = 126.9,
-            )
-        }
     }
 
     private class FakeSpotApiService(
@@ -386,82 +231,6 @@ class CollectionSpotRepositoryImplTest {
                                     spotNm = "재활용센터",
                                     addrBase = "서울특별시 구로구 구로동",
                                     addrDtl = "센터 입구",
-                                ),
-                            ),
-                        ),
-                    ),
-                ),
-            )
-        }
-
-        fun createDuplicateAddressResponse(): SpotResponseDto {
-            return SpotResponseDto(
-                response = SpotResponseBodyDto(
-                    header = SpotHeaderDto(
-                        resultCode = "00",
-                        resultMsg = "NORMAL SERVICE",
-                    ),
-                    body = SpotBodyDto(
-                        items = SpotItemsDto(
-                            item = listOf(
-                                SpotItemDto(
-                                    spotNm = "폐건전지 수거함",
-                                    addrBase = "서울특별시 영등포구 문래동",
-                                    addrDtl = "주민센터 앞",
-                                ),
-                                SpotItemDto(
-                                    spotNm = "폐건전지 수거함 2",
-                                    addrBase = " 서울특별시 영등포구 문래동 ",
-                                    addrDtl = "학교 앞",
-                                ),
-                                SpotItemDto(
-                                    spotNm = "재활용센터",
-                                    addrBase = "서울특별시 구로구 구로동",
-                                    addrDtl = "센터 입구",
-                                ),
-                            ),
-                        ),
-                    ),
-                ),
-            )
-        }
-
-        fun createParenthesizedRoadAddressResponse(): SpotResponseDto {
-            return SpotResponseDto(
-                response = SpotResponseBodyDto(
-                    header = SpotHeaderDto(
-                        resultCode = "00",
-                        resultMsg = "NORMAL SERVICE",
-                    ),
-                    body = SpotBodyDto(
-                        items = SpotItemsDto(
-                            item = listOf(
-                                SpotItemDto(
-                                    spotNm = "종량제봉투 판매소",
-                                    addrBase = "서울특별시 성동구 행당로 35 (금호동1가, 금북빌딩)",
-                                    addrDtl = "103호, 105호",
-                                ),
-                            ),
-                        ),
-                    ),
-                ),
-            )
-        }
-
-        fun createBlankAddressResponse(): SpotResponseDto {
-            return SpotResponseDto(
-                response = SpotResponseBodyDto(
-                    header = SpotHeaderDto(
-                        resultCode = "00",
-                        resultMsg = "NORMAL SERVICE",
-                    ),
-                    body = SpotBodyDto(
-                        items = SpotItemsDto(
-                            item = listOf(
-                                SpotItemDto(
-                                    spotNm = "주소 없는 수거함",
-                                    addrBase = " ",
-                                    addrDtl = "상세 위치",
                                 ),
                             ),
                         ),
