+++ domain/src/main/java/com/team/yeogibeoryeo/domain/spot/repository/CollectionSpotRepository.kt
@@ -7,41 +7,14 @@ import com.team.yeogibeoryeo.domain.spot.model.Coordinate
 
 interface CollectionSpotRepository {
 
-    suspend fun searchByKeyword(
-        keyword: String,
-        types: Set<CollectionSpotType> = emptySet()
-    ): List<CollectionSpot>
-
-    suspend fun searchByKeywordResult(
-        keyword: String,
-        types: Set<CollectionSpotType> = emptySet()
-    ): CollectionSpotSearchResult {
-        return CollectionSpotSearchResult(
-            spots = searchByKeyword(
-                keyword = keyword,
-                types = types,
-            ),
-        )
-    }
-
-    suspend fun searchByKeywordResultWithoutCoordinates(
+    suspend fun searchRawByKeyword(
         keyword: String,
         types: Set<CollectionSpotType> = emptySet()
     ): CollectionSpotSearchResult
 
-    suspend fun searchByLocation(
+    suspend fun searchRawByLocation(
         coordinate: Coordinate,
         radiusMeter: Int,
         types: Set<CollectionSpotType> = emptySet()
     ): List<CollectionSpot>
-
-    suspend fun geocodeSpots(
-        spots: List<CollectionSpot>
-    ): List<CollectionSpot> {
-        return spots.map { spot -> geocodeSpot(spot) }
-    }
-
-    suspend fun geocodeSpot(
-        spot: CollectionSpot
-    ): CollectionSpot
 }
