+++ domain/src/main/java/com/team/yeogibeoryeo/domain/spot/usecase/GeocodeCollectionSpotUseCase.kt
@@ -1,11 +1,11 @@
 package com.team.yeogibeoryeo.domain.spot.usecase
 
 import com.team.yeogibeoryeo.domain.spot.model.CollectionSpot
-import com.team.yeogibeoryeo.domain.spot.repository.CollectionSpotRepository
+import com.team.yeogibeoryeo.domain.spot.repository.CollectionSpotGeocodingRepository
 import javax.inject.Inject
 
 class GeocodeCollectionSpotUseCase @Inject constructor(
-    private val repository: CollectionSpotRepository
+    private val repository: CollectionSpotGeocodingRepository
 ) {
     suspend operator fun invoke(
         spot: CollectionSpot
